backend/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "docplatform-backend"
version = "1.0.0"
description = "AI-Powered Document Generation Platform Backend"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "DocPlatform Team", email = "team@docplatform.com"},
]
keywords = ["fastapi", "document-generation", "ai", "salesforce"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: FastAPI",
]

dependencies = [
    # Core Framework
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    
    # Database
    "sqlalchemy>=2.0.23",
    "alembic>=1.12.1",
    "asyncpg>=0.29.0",
    
    # Background Jobs
    "celery>=5.3.4,<6.0.0",
    "redis>=5.0.1,<6.0.0",
    
    # Document Processing
    "python-docx>=0.8.11,<1.0.0",
    "PyPDF2>=3.0.1,<4.0.0",
    "reportlab>=4.0.7,<5.0.0",
    "openpyxl>=3.1.2,<4.0.0",
    "python-multipart>=0.0.6",
    "Pillow>=10.1.0,<11.0.0",
    "weasyprint>=60.2,<61.0",
    "mammoth>=1.6.0,<2.0.0",
    
    # AI Integration
    "google-generativeai>=0.8.0",
    "openai>=1.3.0,<2.0.0",
    "langchain>=0.0.340,<1.0.0",
    "tiktoken>=0.5.1,<1.0.0",
    "chromadb>=0.4.18,<1.0.0",
    
    # PyTorch & ML
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "torchaudio>=2.1.0",
    "transformers>=4.35.0",
    "datasets>=2.14.0",
    "tokenizers>=0.15.0",
    "sentence-transformers>=2.2.2",
    
    # Salesforce Integration
    "simple-salesforce>=1.12.4,<2.0.0",
    "requests>=2.31.0",
    
    # Storage & Cloud
    "boto3>=1.34.0,<2.0.0",
    "botocore>=1.34.0,<2.0.0",
    
    # Google OAuth
    "google-auth>=2.25.0",
    "google-auth-oauthlib>=1.1.0",
    "google-auth-httplib2>=0.2.0",
    
    # Security & Auth
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    
    # Utilities
    "httpx>=0.25.2,<1.0.0",
    "aiofiles>=23.2.1,<24.0.0",
    "python-dotenv>=1.0.0",
    "structlog>=23.2.0",
    "rich>=13.7.0,<14.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "black>=23.11.0",
    "isort>=5.12.0",
    "mypy>=1.7.1",
    "flake8>=6.1.0,<7.0.0",
    "pre-commit>=3.6.0,<4.0.0",
    "watchdog>=3.0.0,<4.0.0",
    "ipython>=8.17.2,<9.0.0",
    "jupyter>=1.0.0,<2.0.0",
]

production = [
    "gunicorn>=21.2.0,<22.0.0",
    "sentry-sdk[fastapi]>=1.38.0,<2.0.0",
]

[project.urls]
Documentation = "https://docs.docplatform.com"
Repository = "https://github.com/docplatform/docplatform"
Issues = "https://github.com/docplatform/docplatform/issues"

[project.scripts]
docplatform = "app.main:app"

[tool.hatch.version]
path = "app/__about__.py"

[tool.hatch.envs.default]
dependencies = [
    "coverage[toml]>=6.5",
    "pytest",
]

[tool.hatch.envs.default.scripts]
test = "pytest {args:tests}"
test-cov = "coverage run -m pytest {args:tests}"
cov-report = [
    "- coverage combine",
    "coverage report",
]
cov-html = [
    "cov-report",
    "coverage html",
]

[tool.black]
target-version = ["py311"]
line-length = 88
skip-string-normalization = true

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88

[tool.mypy]
strict = true
warn_unreachable = true
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--disable-warnings",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source_pkgs = ["app", "tests"]
branch = true
parallel = true
omit = [
    "app/__about__.py",
]

[tool.coverage.paths]
docplatform = ["app", "*/docplatform-backend/app"]
tests = ["tests", "*/docplatform-backend/tests"]

[tool.coverage.report]
exclude_lines = [
    "no cov",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
] 